networking: refactor implementation in a common networking.c
Move the implementation the actual implementation through the net_ops structure
in a common file.

This is a refactoring step to allow linking both implementation at
the same time.
